pyproject.toml
[project]
name = "rail-harness"
version = "0.1.0"
description = "Rail harness control plane with a Python Actor Runtime."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents==0.8.4",
    "pydantic>=2.12,<3",
    "pyyaml>=6.0.2,<7",
]

[dependency-groups]
dev = [
    "mypy>=1.15,<2",
    "pytest>=8.3,<9",
    "ruff>=0.14,<0.15",
    "types-PyYAML>=6.0,<7",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
